- обновление по raylib от Jul 18, 2021

+++ Code/Game/Player.cpp
@@ -3,8 +3,8 @@
 //-----------------------------------------------------------------------------
 bool Party::Move(const Vector3& pos) noexcept
 {
-	m_position.x = (float)pos.x;
-	m_position.y = (float)pos.z;
+	m_position.x = (int)pos.x;
+	m_position.y = (int)pos.z;
 	printf("PP = %d %d\n", m_position.x, m_position.y);
 
 	if (m_oldPosition.x != m_position.x || m_oldPosition.y != m_position.y)
